infiniteObjectList 반환하는 hook 완성, 사용자 세팅 페이지에 적용
