+++ src/app/api/user/route.ts
@@ -26,6 +26,7 @@ export const GET = async () => {
     return new NextResponse(
       JSON.stringify({
         users: Items,
+        isLast: !LastEvaluatedKey,
         LastEvaluatedKey,
       }),
       {
